telegram_bot/requirements.txt
aiodns==3.0.0
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==22.2.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
dataclasses-json==0.6.0
free-proxy==1.1.1
frozenlist==1.4.0
idna==3.4
lxml==4.9.3
magic-filter==1.0.11
marshmallow==3.20.1
maxminddb==2.4.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
packaging==23.1
proxybroker @ git+https://github.com/bluet/proxybroker2.git@19e57ba6333a965412eebe50c10aba9321815209
pycares==4.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.5
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.5
yarl==1.9.2
